+++ ci/travis.sh
@@ -8,6 +8,7 @@ case "$1" in
     case "$2" in
       audit) ci/travis/audit-install.sh;;
       clippy) ci/travis/clippy-install.sh;;
+      fmt) ci/travis/fmt-install.sh;;
       test) ;;
 
       *)
@@ -20,6 +21,7 @@ case "$1" in
     case "$2" in
       audit) ci/travis/audit-script.sh;;
       clippy) ci/travis/clippy-script.sh;;
+      fmt) ci/travis/fmt-script.sh;;
       test) ci/travis/test-script.sh;;
 
       *)