+++ replictl/src/error.rs
@@ -77,6 +77,18 @@ pub enum ErrorKind {
     #[fail(display = "I/O error on file {}", _0)]
     Io(String),
 
+    #[fail(display = "could not list non-blocking locks from coordinator")]
+    NBLockList,
+
+    #[fail(display = "could not lookup non-blocking lock '{}'", _0)]
+    NBLockLookup(String),
+
+    #[fail(display = "could not lookup current owner of non-blocking lock '{}'", _0)]
+    NBLockOwnerLookup(String),
+
+    #[fail(display = "could not force-release non-blocking lock '{}'", _0)]
+    NBLockRelease(String),
+
     #[fail(display = "need a command to run for '{}'", _0)]
     NoCommand(&'static str),