+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 ### Added
+- Add `rustfmt` to CI and move in that direction
 - Add coordinator version to `replictl versions`.
 - Add queue system version to `replictl versions`.
 - Contributing Guidelines.